refactoring word tokenizer
