+++ lib/classifiers/classifier.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const EventEmitter = require('events').EventEmitter;
-const tagger = require('../word_sent');
+const tagger = require('../word_tokenizer');
 const stopwords = require('../util/stopwords');
 
 class Classifier extends EventEmitter {
