+++ lib/pos_tag/index.js
@@ -1,9 +1,8 @@
 'use strict';
 const path = require('path');
 const crfsuite = require('crfsuite');
-const tokenizer = require('../tokenizer');
-const word_sent = require('../word_sent');
-const fe = require('../features');
+const tokenizer = require('../word_tokenizer');
+const features = require('../features');
 
 const logger = require('../logger')('POSTag');
 
@@ -42,7 +41,7 @@ class POSTag {
     }
 
     tag(text) {
-        let words = word_sent.tag(text);
+        let words = tokenizer.tag(text);
         let tokens = words.map((token) => {
             return [token, 'X']
         });
@@ -54,7 +53,7 @@ class POSTag {
 
     transform(tokens) {
         let template = this.template;
-        return tokens.map((token, i) => fe.word2features(tokens, i, template));
+        return tokens.map((token, i) => features.word2features(tokens, i, template));
     }
 }
