+++ lib/vntk.js
@@ -18,11 +18,11 @@ exports.tokenizer = () => require('./tokenizer');
  * Word Segmentation
  * @param {String} modelFileName new custom model
  */
-exports.wordSent = (modelFileName) => {
+exports.wordTokenizer = (modelFileName) => {
     if(modelFileName && fs.existsSync(modelFileName)) {
-        return require('./word_sent').newModel(modelFileName)
+        return require('./word_tokenizer').newModel(modelFileName)
     } else  {
-        return require('./word_sent')
+        return require('./word_tokenizer')
     }
 }
