+++ lib/word_tokenizer/index.js
@@ -4,9 +4,9 @@ const crfsuite = require('crfsuite');
 const tokenizer = require('../tokenizer');
 const fe = require('../features');
 
-const logger = require('../logger')('WordSent');
+const logger = require('../logger')('WordTokenizer');
 
-class WordSent {
+class WordTokenizer {
 
     constructor(fn) {
         this.tagger = crfsuite.Tagger();
@@ -24,7 +24,7 @@ class WordSent {
      * @param {String} fn filename
      */
     newModel(fn) {
-        return new WordSent(fn);
+        return new WordTokenizer(fn);
     }
 
     get template() {
@@ -100,4 +100,4 @@ class WordSent {
     }
 }
 
-module.exports = new WordSent();
+module.exports = new WordTokenizer();
